Moved HTTP API methods into topic specific files
